+++ component/backend/Toolbar/Toolbar.php
@@ -26,7 +26,6 @@ class Toolbar extends \FOF30\Toolbar\Toolbar
 			'ControlPanel',
 			'COM_AKEEBASUBS_MAINMENU_SETUP'    => array(
 				'Levels',
-				'LevelGroups',
 				'Relations',
 				'Upgrades',
 				'TaxConfig',