+++ component/backend/ViewTemplates/Levels/form.priceduration.blade.php
@@ -41,18 +41,3 @@ defined('_JEXEC') or die();
     </label>
     @jhtml('FEFHelper.select.booleanswitch', 'forever', $item->forever)
 </div>
-
-<div class="akeeba-form-group">
-    <label for="akeebasubs_levelgroup_id">
-        @lang('COM_AKEEBASUBS_LEVELS_FIELD_LEVELGROUP')
-    </label>
-    <?php echo \FOF30\Utils\FEFHelper\BrowseView::modelSelect('akeebasubs_levelgroup_id', 'LevelGroups',
-		$item->akeebasubs_levelgroup_id, [
-			'translate'      => false,
-			'apply_access'   => false,
-			'key_field'      => "akeebasubs_levelgroup_id",
-			'value_field'    => "title",
-			'none'           => "COM_AKEEBASUBS_SELECT_LEVELGROUP",
-			'fof.autosubmit' => false
-		]) ?>
-</div>